requirements.txt
Django~=6.0.4
environs~=15.0.1
dj-database-url~=3.1.2
django-tailwind~=4.4.2
whitenoise~=6.12.0
django-two-factor-auth~=1.18.1
django-unfold~=0.92.0
django-ninja~=1.5.2
httpx~=0.28.1
